debundle/lowering: extract imports_runtime
BAZEL_TEST_INVOCATIONS=none: file move.
